Safari-611.1.10/LayoutTests/accessibility/mac/button-shouldnot-have-axvalue-expected.txt


This tests that a button should not have AXRequired and should not have AXValue returned.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS button.isAttributeSupported('AXValue') is false
PASS button.isAttributeSupported('AXRequired') is false
PASS successfullyParsed is true

TEST COMPLETE
